music recommendation/requirements.txt
google-generativeai
pandas
numpy
faiss-cpu
pyyaml
Flask
sentence-transformers
torch